gui: split formatting into dedicated file
